Support dynamic lists for Hydra configs
While specifying a true list in the config files, such as a specific subset
of files to download from The Pile, the output will be generated in another
launcher script that is incompatible with Hydra. In order for the list to
be interpreted correctly, it needs to be converted to a string
representation of a list with no spaces.

Signed-Off-By: Robert Clark <roclark@nvidia.com>